# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank REST API
  description: API управления банковскими картами (карты, переводы, JWT, роли ADMIN/USER)
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Локальный dev

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        username: { type: string }
        role: { type: string, enum: [ADMIN, USER] }
    CardResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        maskedNumber: { type: string, example: "**** **** **** 1234" }
        holderName: { type: string }
        expiryDate: { type: string, format: date }
        status: { type: string, enum: [ACTIVE, BLOCKED, EXPIRED] }
        balance: { type: number, format: decimal }
        createdAt: { type: string, format: date-time }
    CardCreateRequest:
      type: object
      required: [number, holderName, expiryDate]
      properties:
        number: { type: string, description: "16-19 цифр" }
        holderName: { type: string }
        expiryDate: { type: string, format: date }
    TransferRequest:
      type: object
      required: [fromCardId, toCardId, amountCents]
      properties:
        fromCardId: { type: integer, format: int64 }
        toCardId: { type: integer, format: int64 }
        amountCents: { type: integer, minimum: 1 }
    UserCreateRequest:
      type: object
      required: [username, password, role]
      properties:
        username: { type: string }
        password: { type: string, format: password }
        role: { type: string, enum: [ADMIN, USER] }
    ApiError:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        path: { type: string }
        details: { type: array, items: { type: string } }

paths:
  /api/auth/login:
    post:
      summary: Логин
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthRequest' }
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Неверные креды
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }

  /api/cards:
    get:
      summary: Список карт (свои для USER, все для ADMIN)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [ACTIVE, BLOCKED, EXPIRED] }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Страница карт
    post:
      summary: Создать карту
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CardCreateRequest' }
      responses:
        '201':
          description: Создана
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CardResponse' }

  /api/cards/{id}:
    get:
      summary: Карта по id
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CardResponse' }
    patch:
      summary: Изменить статус (USER — BLOCKED, ADMIN — любой)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          required: true
          schema: { type: string, enum: [ACTIVE, BLOCKED, EXPIRED] }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CardResponse' }
    delete:
      summary: Удалить карту (только ADMIN)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: Удалено }

  /api/transfers:
    post:
      summary: Перевод между своими картами
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransferRequest' }
      responses:
        '201':
          description: Перевод выполнен
        '400':
          description: Ошибка валидации/баланса

  /api/admin/users:
    post:
      summary: Создать пользователя (ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreateRequest' }
      responses:
        '201':
          description: Пользователь создан