# src/main/java/com/example/bankcards/dto/UserCreateRequest.java
package com.example.bankcards.dto;

import com.example.bankcards.entity.Role;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class UserCreateRequest {
    @NotBlank
    @Size(min = 1, max = 255)
    private String username;
    @NotBlank
    @Size(min = 4, max = 100)
    private String password;
    @NotNull
    private Role role;
}